requirements.txt
streamlit
pymupdf
langchain-google-genai
langchain-groq
pandas
pillow
python-dotenv
langchain-core